requirements.txt
Flask
Gunicorn
psycopg2-binary==2.9.11
